+++ README.md
@@ -1,4 +1,4 @@
-# Spatial aggregation
+# Gregor
 
 ## Installation
