+++ src/gregor/aggregate.py
