+++ src/gregor/raster.py
