--- test/test.py
@@ -1,6 +1,0 @@
-import src.spatial_aggregation as spatial_aggregation
-
-
-def test_get_uniform_proxy():
-    
-    spatial_aggregation.get_uniform_proxy(spatial_units, raster_resolution)
